Support for two types of vaults. Support for policies using uvm-compliant status. Tested with Rego policies.

+++ inference/models/generatefs.sh
@@ -12,15 +12,18 @@ echo Encrypting $dataPath with key $keyFilePath and generating $encryptedImage
 deviceName=cryptdevice1
 deviceNamePath="/dev/mapper/$deviceName"
 
-echo "[!] Generating keyfile..."
-
-dd if=/dev/random of="$keyFilePath" count=1 bs=32
+if [ -f "$keyFilePath" ]; then
+    echo "keyfile exists"
+else
+    echo "[!] Generating keyfile..."
+    dd if=/dev/random of="$keyFilePath" count=1 bs=32
+fi
 
 echo "Key in hex string format"
 
 python hexstring.py $keyFilePath
 
-truncate -s 64 "$keyFilePath"
+truncate -s 32 "$keyFilePath"
 
 echo "[!] Creating encrypted image..."
